Refactored button-related files
