+++ libs/app-theme/ui/src/lib/button/button-primary.styles.tsx
@@ -1,0 +1,80 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import styled, { css } from "styled-components"
+import { Link } from "react-router"
+import {
+  baseStyles,
+  ButtonIcon,
+  StyledButtonProps,
+  StyledLinkProps,
+} from "./button-base.styles"
+
+const primaryDefaultStyles = css`
+  background-color: ${({ theme }) => theme.app.color.grey1};
+  transition:
+    color 0.15s ease-in-out,
+    background-color 0.15s ease-in-out,
+    border-color 0.15s ease-in-out;
+  font-size: ${({ theme }) => theme.app.fontSize.buttonText};
+  line-height: ${({ theme }) => theme.app.lineHeight.buttonText};
+  color: ${({ theme }) => theme.app.color.white};
+  height: 4rem;
+
+  ${ButtonIcon} {
+    width: 2rem;
+    height: 2rem;
+    margin-right: ${({ theme }) => theme.app.space.xs};
+  }
+`
+
+const primaryHoverStyles = css`
+  background-color: ${({ theme }) => theme.app.color.black};
+`
+
+const primaryActiveStyles = css`
+  background-color: ${({ theme }) => theme.app.color.grey1};
+`
+
+const primaryDisabledStyles = css`
+  background-color: ${({ theme }) => theme.app.color.grey4};
+  color: ${({ theme }) => theme.app.color.grey2};
+  cursor: not-allowed;
+`
+
+export const PrimaryButtonComponent = styled.button<StyledButtonProps>`
+  ${baseStyles};
+  ${primaryDefaultStyles};
+
+  &:hover {
+    ${primaryHoverStyles};
+  }
+  &:active {
+    ${primaryActiveStyles};
+  }
+  &:disabled {
+    ${primaryDisabledStyles};
+  }
+`
+
+export const PrimaryNavigationComponent = styled(Link)<StyledLinkProps>`
+  ${baseStyles};
+  ${primaryDefaultStyles};
+
+  ${({ $disabled }) =>
+    $disabled
+      ? css`
+          ${primaryDisabledStyles};
+        `
+      : css`
+          &:hover {
+            ${primaryHoverStyles};
+          }
+
+          &:active {
+            ${primaryActiveStyles};
+          }
+        `};
+`
